demo/requirements.txt
st-social-media-links
st_supabase_connection
streamlit_extras
